2. Тесты для задачи "идеальный вес" [#505260 #529593] исправил!!

+++ src/test/java/ru/job4j/ConverterTest.java
@@ -2,13 +2,15 @@ package ru.job4j;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
 class ConverterTest {
     @Test
     void whenConvert140RblThen2Euro() {
         float input = 140;
         float expected = 1.5555555f;
-        float output = Converter.rubleToEuro(input);
+        float output = (float) Converter.rubleToEuro(input);
         float value = 0.0001f;
         assertThat(output).isEqualTo(expected, Assertions.withPrecision(value));
     }
@@ -17,7 +19,7 @@ class ConverterTest {
     void whenConvert180RbThen3Euro() {
         float input = 180;
         float expected = 3;
-        float output = Converter.rubleToDollar(input);
+        float output = (float) Converter.rubleToDollar(input);
         float value = 0.0001f;
         assertThat(output).isEqualTo(expected, Assertions.withPrecision(value));
     }
